requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
selenium==4.15.2
webdriver-manager==4.0.1
setuptools>=65.5.1
aiohttp==3.9.1
asyncio==3.4.3
asgiref==3.7.2
uvicorn==0.24.0
apify-client
bcrypt==4.1.2
